Added sample JSON and script for generating sample files
The 001.json is the one that is used as an example in the lab in order to test the sites functionality you do the post request with thid json and the first comment appears on the main page.
generate_feedback is intended to allow you to generate your own feedback files either from a CSV or manually but it is as for this moment not completed.