Added this script for rewriting car_sales.json
This script reads the json converts it to a dictionary.
If the add option is selected the script reads the current state of the json and creates a list of dictionaries based on it prompts the user to add every key-value pair needed in order to create a new element in the list.
If the new option is selected the script prompts the user for the key-value pairs of the new elements and adds them in the empty dictionary.
Both branches rewrite car_sales.json as appending was causing issues.